this and that. In particular, the HasXxxId interfaces were at two different places; I consolidated them.  In freight, making code shorter.
