+++ contribs/dvrp/src/main/java/org/matsim/contrib/dvrp/passenger/AbstractPassengerEvent.java
@@ -27,7 +27,7 @@ import org.matsim.api.core.v01.events.Event;
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.contrib.dvrp.fleet.DvrpVehicle;
 import org.matsim.contrib.dvrp.optimizer.Request;
-import org.matsim.core.api.internal.HasPersonId;
+import org.matsim.api.core.v01.events.HasPersonId;
 
 /**
  * @author Michal Maciejewski (michalm)
