+++ contribs/dvrp/src/main/java/org/matsim/contrib/dvrp/vrpagent/AbstractTaskEvent.java
@@ -29,7 +29,7 @@ import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.contrib.dvrp.fleet.DvrpVehicle;
 import org.matsim.contrib.dvrp.schedule.Task.TaskType;
-import org.matsim.core.api.internal.HasPersonId;
+import org.matsim.api.core.v01.events.HasPersonId;
 
 /**
  * @author Michal Maciejewski (michalm)
