+++ contribs/noise/src/main/java/org/matsim/contrib/noise/personLinkMoneyEvents/PersonLinkMoneyEvent.java
@@ -26,7 +26,7 @@ import org.matsim.api.core.v01.events.Event;
 import org.matsim.api.core.v01.events.PersonMoneyEvent;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.population.Person;
-import org.matsim.core.api.internal.HasPersonId;
+import org.matsim.api.core.v01.events.HasPersonId;
 
 /**
  * The idea is to have a standard events format to be used by various external cost pricing approaches.
