+++ matsim/src/main/java/org/matsim/api/core/v01/events/LinkLeaveEvent.java
@@ -20,19 +20,18 @@
 
 package org.matsim.api.core.v01.events;
 
-import java.util.Map;
-
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.vehicles.Vehicle;
 
-public class LinkLeaveEvent extends Event implements HasLinkId {
+import java.util.Map;
+
+public class LinkLeaveEvent extends Event implements HasLinkId, HasVehicleId{
 
 	public static final String EVENT_TYPE = "left link";
 	public static final String ATTRIBUTE_LINK = "link";
-	public static final String ATTRIBUTE_VEHICLE = "vehicle";
-	
+
 	private final Id<Link> linkId;
 	private final Id<Vehicle> vehicleId;
