+++ matsim/src/main/java/org/matsim/api/core/v01/events/PersonEntersVehicleEvent.java
@@ -24,7 +24,6 @@ import java.util.Map;
 
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.population.Person;
-import org.matsim.core.api.internal.HasPersonId;
 import org.matsim.vehicles.Vehicle;
 
 /**
