+++ matsim/src/main/java/org/matsim/api/core/v01/events/PersonMoneyEvent.java
@@ -24,7 +24,6 @@ import java.util.Map;
 
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.population.Person;
-import org.matsim.core.api.internal.HasPersonId;
 
 /**
  * This event specifies that an agent has gained (or paid) some money.
