+++ matsim/src/main/java/org/matsim/api/core/v01/events/PersonScoreEvent.java
@@ -19,9 +19,7 @@
 package org.matsim.api.core.v01.events;
 
 import org.matsim.api.core.v01.Id;
-import org.matsim.api.core.v01.events.Event;
 import org.matsim.api.core.v01.population.Person;
-import org.matsim.core.api.internal.HasPersonId;
 
 import java.util.Map;
