+++ matsim/src/main/java/org/matsim/api/core/v01/events/PersonStuckEvent.java
@@ -25,7 +25,6 @@ import java.util.Map;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.population.Person;
-import org.matsim.core.api.internal.HasPersonId;
 
 public class PersonStuckEvent extends Event implements HasPersonId {
