+++ matsim/src/main/java/org/matsim/api/core/v01/events/VehicleEntersTrafficEvent.java
@@ -25,8 +25,6 @@ import java.util.Map;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.population.Person;
-import org.matsim.core.api.internal.HasPersonId;
-import org.matsim.core.api.internal.HasVehicleId;
 import org.matsim.vehicles.Vehicle;
 
 public class VehicleEntersTrafficEvent extends Event implements HasPersonId, HasLinkId, HasVehicleId{
