+++ matsim/src/main/java/org/matsim/core/api/experimental/events/AgentWaitingForPtEvent.java
@@ -24,7 +24,7 @@ import java.util.Map;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.events.Event;
 import org.matsim.api.core.v01.population.Person;
-import org.matsim.core.api.internal.HasPersonId;
+import org.matsim.api.core.v01.events.HasPersonId;
 import org.matsim.pt.transitSchedule.api.TransitStopFacility;
 
 /**
